local_code/requirements.txt
packaging==25.0
paho-mqtt==2.1.0
python-can==4.6.1
typing_extensions==4.15.0
wrapt==1.17.3
pymongo==4.10.1
python-dotenv==1.0.1
